% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_lm.R
\name{build_lm}
\alias{build_lm}
\title{Fit a Bayesian glm in nimble where the model was first defined and fitted using stan_glm()}
\usage{
build_lm(stanobj)
}
\arguments{
\item{stanobj}{A stanfit object created from a call to stan_glm()}
}
\value{
A stanfit object which is a copy of the stanjob passed and retrofitted
  with results from nimble
}
\description{
Core purpose of this function is to compare posterior estimates between rstanarm and nimble
}
