% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_glm.R
\name{nimble_glm}
\alias{nimble_glm}
\title{Fit a Bayesian glm in nimble where the model was first defined and fitted using stan_glm()}
\usage{
nimble_glm(
  stanobj,
  seed = c(12345),
  warmup = 100,
  iter = 1000,
  thin = 1,
  chains = 1
)
}
\arguments{
\item{stanobj}{A stanfit object created from a call to stan_glm()}

\item{seed}{A vector of integers same length as number of chains}

\item{warmup}{A single integer for the burn-in in each chain}

\item{iter}{A single integer for total number of mcmc interations}

\item{chains}{A single integer for number of independent mcmc chains to be run
each end of x before the mean is computed}
}
\value{
A stanfit object which is a copy of the stanjob passed and retrofitted
  with results from nimble
}
\description{
Core purpose of this function is to compare posterior estimates between rstanarm and nimble
}
\examples{
nimble_glm(mystanobj,c(12345,12345), warmup=1000, iter=100000, thin=1, chains=2)



}
